frontend/requirements.txt
streamlit==1.45.1
requests==2.32.3
pandas==2.3.0
plotly==6.1.2
python-dotenv
